backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
paddlepaddle
paddleocr>=3.3.2
pillow>=11.1.0
opencv-python>=4.11.0
